TDD/TDD_With_JUnit5_And_Mockito/code/8-implementing-air-conditioning/flight-mgmt-app/flight-mgmt-app/src/main/java/com/shshetudev/flightmgmtapp/airconditioning/AirConditioningSystem.java
package com.shshetudev.flightmgmtapp.airconditioning;

import org.junit.jupiter.api.extension.ExtendWith;

public class AirConditioningSystem {
    private Thermometer thermometer;
    private double temperatureThreshold;
    private boolean open;

    public AirConditioningSystem() {
        open = false;
    }

    // Initially it was empty
    public void checkAirConditioningSystem() {
        this.open = thermometer.getTemperature() >= temperatureThreshold;
    }

    public boolean isOpen() {
        return open;
    }

    public void setThermometer(Thermometer thermometer) {
        this.thermometer = thermometer;
    }

    public void setTemperatureThreshold(double temperatureThreshold) {
        this.temperatureThreshold = temperatureThreshold;
    }
}
